+++ RELEASE_NOTES.md
@@ -33,6 +33,12 @@ Documentation
 
 * FIXED: Missing `loops` library API documentation.
 
+IDEs, text editors, and syntax highlighters support
+---------------------------------------------------
+
+* UPDATED: Syntax highlighting test files to test representing a single quote
+inside a quoted atom using an escape sequence.
+
 
 3.27.0 - June 18, 2019
 ======================